+++ grid/Grid.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31702.278
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Grid", "Grid\Grid.csproj", "{8B11A421-D335-4332-89A0-21C09009C8D1}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{8B11A421-D335-4332-89A0-21C09009C8D1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8B11A421-D335-4332-89A0-21C09009C8D1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8B11A421-D335-4332-89A0-21C09009C8D1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8B11A421-D335-4332-89A0-21C09009C8D1}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {71DDC4DF-7567-4473-A4AA-0C974517A437}
+	EndGlobalSection
+EndGlobal
