+++ grid/Grid/MainWindow.xaml
@@ -1,0 +1,136 @@
+﻿<Window x:Class="Grid.MainWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:Grid"
+
+        mc:Ignorable="d"
+        Title="MainWindow" Height="450" Width="800">
+    <Grid>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="20*"/>
+            <ColumnDefinition Width="60*"/>
+        </Grid.ColumnDefinitions>
+        <ProgressBar Minimum="0" Maximum="100" Name="pbStatus" />
+        <TabControl Name="Emotions" Margin="10,5,-3,5" Grid.Column="1">
+            <TabItem Header="neutral">
+                <Grid Background="#FFE5E5E5">
+                    <ListBox Name="neutral" Margin="0,0,250,-1" >
+                        <ListBox.ItemTemplate>
+                            <DataTemplate>
+                                <StackPanel Orientation="Horizontal">
+                                    <Image Source="{Binding ImgPath}" Width="100" Height="100"></Image>
+                                </StackPanel>
+                            </DataTemplate>
+                        </ListBox.ItemTemplate>
+                    </ListBox>
+                    <ListBox Margin="348,16,10,15" ItemsSource="{Binding ElementName=neutral, Path=SelectedItem.Emos}"/>
+                </Grid>
+            </TabItem>
+            <TabItem Header="happiness">
+                <Grid Background="#FFE5E5E5">
+                    <ListBox Name="happiness" Margin="0,0,250,0" ItemsSource="{Binding}">
+                        <ListBox.ItemTemplate>
+                            <DataTemplate>
+                                <StackPanel Orientation="Horizontal">
+                                    <Image Source="{Binding ImgPath}"></Image>
+                                </StackPanel>
+                            </DataTemplate>
+                        </ListBox.ItemTemplate>
+                    </ListBox>
+                    <ListBox Margin="348,16,10,15" ItemsSource="{Binding ElementName=happiness, Path=SelectedItem.Emos}"/>
+                </Grid>
+            </TabItem>
+            <TabItem Header="surprise">
+                <Grid Background="#FFE5E5E5">
+                    <ListBox Name="surprise" Margin="0,0,250,0" ItemsSource="{Binding}">
+                        <ListBox.ItemTemplate>
+                            <DataTemplate>
+                                <StackPanel Orientation="Horizontal">
+                                    <Image Source="{Binding ImgPath}"></Image>
+                                </StackPanel>
+                            </DataTemplate>
+                        </ListBox.ItemTemplate>
+                    </ListBox>
+                    <ListBox Margin="348,16,10,15" ItemsSource="{Binding ElementName=surprise, Path=SelectedItem.Emos}"/>
+                </Grid>
+            </TabItem>
+            <TabItem Header="sadness">
+                <Grid Background="#FFE5E5E5">
+                    <ListBox Name="sadness" Margin="0,0,250,0" ItemsSource="{Binding}">
+                        <ListBox.ItemTemplate>
+                            <DataTemplate>
+                                <StackPanel Orientation="Horizontal">
+                                    <Image Source="{Binding ImgPath}"></Image>
+                                </StackPanel>
+                            </DataTemplate>
+                        </ListBox.ItemTemplate>
+                    </ListBox>
+                    <ListBox Margin="348,16,10,15" ItemsSource="{Binding ElementName=sadness, Path=SelectedItem.Emos}"/>
+                </Grid>
+            </TabItem>
+            <TabItem Header="anger">
+                <Grid Background="#FFE5E5E5">
+                    <ListBox Name="anger" Margin="0,0,250,0" ItemsSource="{Binding}">
+                        <ListBox.ItemTemplate>
+                            <DataTemplate>
+                                <StackPanel Orientation="Horizontal">
+                                    <Image Source="{Binding ImgPath}"></Image>
+                                </StackPanel>
+                            </DataTemplate>
+                        </ListBox.ItemTemplate>
+                    </ListBox>
+                    <ListBox Margin="348,16,10,15" ItemsSource="{Binding ElementName=anger, Path=SelectedItem.Emos}"/>
+                </Grid>
+            </TabItem>
+            <TabItem Header="disgust">
+                <Grid Background="#FFE5E5E5">
+                    <ListBox Name="disgust" Margin="0,0,250,0" ItemsSource="{Binding}">
+                        <ListBox.ItemTemplate>
+                            <DataTemplate>
+                                <StackPanel Orientation="Horizontal">
+                                    <Image Source="{Binding ImgPath}"></Image>
+                                </StackPanel>
+                            </DataTemplate>
+                        </ListBox.ItemTemplate>
+                    </ListBox>
+                    <ListBox Margin="348,16,10,15" ItemsSource="{Binding ElementName=disgust, Path=SelectedItem.Emos}"/>
+                </Grid>
+            </TabItem>
+            <TabItem Header="fear">
+                <Grid Background="#FFE5E5E5">
+                    <ListBox Name="fear" Margin="0,0,250,0" ItemsSource="{Binding}">
+                        <ListBox.ItemTemplate>
+                            <DataTemplate>
+                                <StackPanel Orientation="Horizontal">
+                                    <Image Source="{Binding ImgPath}"></Image>
+                                </StackPanel>
+                            </DataTemplate>
+                        </ListBox.ItemTemplate>
+                    </ListBox>
+                    <ListBox Margin="348,16,10,15" ItemsSource="{Binding ElementName=fear, Path=SelectedItem.Emos}"/>
+                </Grid>
+            </TabItem>
+            <TabItem Header="contempt">
+                <Grid Background="#FFE5E5E5">
+                    <ListBox Name="contempt" Margin="0,0,250,0" ItemsSource="{Binding}">
+                        <ListBox.ItemTemplate>
+                            <DataTemplate>
+                                <StackPanel Orientation="Horizontal">
+                                    <Image Source="{Binding ImgPath}"></Image>
+                                </StackPanel>
+                            </DataTemplate>
+                        </ListBox.ItemTemplate>
+                    </ListBox>
+                    <ListBox Margin="348,16,10,15" ItemsSource="{Binding ElementName=contempt, Path=SelectedItem.Emos}"/>
+                </Grid>
+            </TabItem>
+        </TabControl>
+        <Button Name ="Folder_button" Content="Choose folder" HorizontalAlignment="Center" Height="43" Margin="0,33,0,0" VerticalAlignment="Top" Width="168" Background="#FFEDFFF1" Click="Button_Click"/>
+        <Button Name="Start_Button" Content="Start" HorizontalAlignment="Center" Height="35" Margin="0,95,0,0" VerticalAlignment="Top" Width="168" Background="#FFD2E6FD" Click="Run"/>
+        <Button Content="Cancel" HorizontalAlignment="Center" Height="35" Margin="0,153,0,0" VerticalAlignment="Top" Width="168" Background="#FFFDD6D2"/>
+        <ListBox Name="Check" HorizontalAlignment="Left" Height="194" Margin="16,210,0,0" VerticalAlignment="Top" Width="157" ItemsSource="{Binding}"/>
+
+    </Grid>
+</Window>
