--- grid/MainWindow.xaml
@@ -1,63 +1,0 @@
-﻿<Window x:Class="WpfApp1.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:WpfApp1"
-        mc:Ignorable="d"
-        Title="MainWindow" Height="450" Width="800">
-    <Grid>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="20*"/>
-            <ColumnDefinition Width="60*"/>
-        </Grid.ColumnDefinitions>
-        <ProgressBar Minimum="0" Maximum="100" Name="pbStatus" />
-        <TabControl Margin="10,5,-3,5" Grid.Column="1">
-            <TabItem Header="neutral">
-                <Grid Background="#FFE5E5E5">
-                    <ListBox Name="neutral" ItemsSource="{Binding}"/>
-                </Grid>
-            </TabItem>
-            <TabItem Header="happiness">
-                <Grid Background="#FFE5E5E5">
-                    <ListBox Name="happiness" ItemsSource="{Binding}"/>
-                </Grid>
-            </TabItem>
-            <TabItem Header="surprise">
-                <Grid Background="#FFE5E5E5">
-                    <ListBox Name="surprise" ItemsSource="{Binding}"/>
-                </Grid>
-            </TabItem>
-            <TabItem Header="sadness">
-                <Grid Background="#FFE5E5E5">
-                    <ListBox Name="sadness" ItemsSource="{Binding}"/>
-                </Grid>
-            </TabItem>
-            <TabItem Header="anger">
-                <Grid Background="#FFE5E5E5">
-                    <ListBox Name="anger" ItemsSource="{Binding}"/>
-                </Grid>
-            </TabItem>
-            <TabItem Header="disgust">
-                <Grid Background="#FFE5E5E5">
-                    <ListBox Name="disgust" ItemsSource="{Binding}"/>
-                </Grid>
-            </TabItem>
-            <TabItem Header="fear">
-                <Grid Background="#FFE5E5E5">
-                    <ListBox Name="fear" ItemsSource="{Binding}"/>
-                </Grid>
-            </TabItem>
-            <TabItem Header="contempt">
-                <Grid Background="#FFE5E5E5">
-                    <ListBox Name="contempt" ItemsSource="{Binding}"/>
-                </Grid>
-            </TabItem>
-        </TabControl>
-        <Button Name ="Folder_button" Content="Choose folder" HorizontalAlignment="Center" Height="43" Margin="0,33,0,0" VerticalAlignment="Top" Width="168" Background="#FFEDFFF1" Click="Button_Click"/>
-        <Button Name="Start_Button" Content="Start" HorizontalAlignment="Center" Height="35" Margin="0,95,0,0" VerticalAlignment="Top" Width="168" Background="#FFD2E6FD" Click="Run"/>
-        <Button Content="Cancel" HorizontalAlignment="Center" Height="35" Margin="0,153,0,0" VerticalAlignment="Top" Width="168" Background="#FFFDD6D2"/>
-        <ListBox Name="Statistics" HorizontalAlignment="Center" Height="207" Margin="0,217,0,0" VerticalAlignment="Top" Width="168" DataContext="{Binding}"/>
-
-    </Grid>
-</Window>
