--- grid/WpfApp1.csproj
@@ -1,23 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>WinExe</OutputType>
-	<TargetFramework>net6.0-windows</TargetFramework>
-    <UseWPF>true</UseWPF>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Reference Include="EmotionFerPlus">
-      <HintPath>D:\Личное\Учеба\GitHub\441_elfimova\lib\obj\Debug\net6.0\EmotionFerPlus.dll</HintPath>
-    </Reference>
-    <Reference Include="Ookii.Dialogs.Wpf">
-      <HintPath>D:\Личное\Учеба\4 курс\Ookii\Ookii.Dialogs.Wpf.dll</HintPath>
-    </Reference>
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="441_elfimova.EFP" Version="0.5.2209.121" />
-    <PackageReference Include="Microsoft.ML.OnnxRuntime" Version="1.13.1" />
-  </ItemGroup>
-
-</Project>
