--- grid/global.json
@@ -1,5 +1,0 @@
-{
-  "sdk": {
-    "version": "6.0.402"
-  }
-}
